Chore: Refactor generators for views, Jbuilder, and scaffold in Securial
- Implemented ViewsGenerator to copy model-related views for customization.
- Created JbuilderGenerator to generate Jbuilder templates for resource representation.
- Added scaffold generator to streamline the creation of models, controllers, views, and specs.
- Introduced middleware for transforming request and response keys.
- Enhanced logging capabilities with a new logger structure and broadcaster.
- Implemented version checker to notify users of available updates.
- Refactored helper modules for better organization and accessibility.
- Updated specs to reflect new module structures and paths.
