+++ .ruby-version
@@ -1,1 +1,1 @@
-ruby-3.4.3
+ruby-3.4.4
