--- lib/securial/auth/_index.rb
@@ -1,3 +1,0 @@
-require_relative "errors"
-require_relative "auth_encoder"
-require_relative "session_creator"
