+++ lib/securial/config.rb
@@ -1,0 +1,10 @@
+require "securial/config/configuration"
+require "securial/config/validation"
+require "securial/config/errors"
+
+module Securial
+  module Config
+    # This module acts as a namespace for configuration-related components.
+    # It requires all key config files: Configuration, Validation, Errors.
+  end
+end
