--- lib/securial/config/_index.rb
@@ -1,3 +1,0 @@
-require_relative "./configuration"
-require_relative "./validation"
-require_relative "./errors"
