+++ lib/securial/generators/factory_bot/model/model_generator.rb
