+++ lib/securial/generators/factory_bot/templates/factory.erb
