+++ lib/securial/generators/install/install_generator.rb
