+++ lib/securial/generators/install/templates/securial_initializer.erb
