+++ lib/securial/generators/install/views_generastor.rb
