+++ lib/securial/generators/jbuilder/jbuilder_generator.rb
