+++ lib/securial/generators/jbuilder/templates/_resource.json.erb
