+++ lib/securial/generators/jbuilder/templates/index.json.erb
