+++ lib/securial/generators/jbuilder/templates/show.json.erb
