+++ lib/securial/generators/scaffold/scaffold_generator.rb
