+++ lib/securial/generators/scaffold/templates/controller.erb
