+++ lib/securial/generators/scaffold/templates/request_spec.erb
