+++ lib/securial/generators/scaffold/templates/routes.erb
