+++ lib/securial/generators/scaffold/templates/routing_spec.erb
