+++ lib/securial/helpers.rb
@@ -1,0 +1,9 @@
+require "securial/helpers/normalizing_helper"
+require "securial/helpers/regex_helper"
+
+module Securial
+  module Helpers
+    # This module acts as a namespace for helper modules.
+    # It requires all helper modules to make them available for consumers.
+  end
+end
