--- lib/securial/helpers/_index.rb
@@ -1,2 +1,0 @@
-require_relative "normalizing_helper"
-require_relative "regex_helper"
