+++ lib/securial/helpers/normalizing_helper.rb
@@ -1,17 +1,19 @@
 module Securial
-  module NormalizingHelper
-    module_function
+  module Helpers
+    module NormalizingHelper
+      module_function
 
-    def normalize_email_address(email)
-      return "" if email.empty?
+      def normalize_email_address(email)
+        return "" if email.empty?
 
-      email.strip.downcase
-    end
+        email.strip.downcase
+      end
 
-    def normalize_role_name(role_name)
-      return "" if role_name.empty?
+      def normalize_role_name(role_name)
+        return "" if role_name.empty?
 
-      role_name.strip.downcase.titleize
+        role_name.strip.downcase.titleize
+      end
     end
   end
 end
