+++ lib/securial/inspectors.rb
@@ -1,0 +1,9 @@
+require "securial/inspectors/route_inspector"
+
+module Securial
+  module Inspectors
+    # This module serves as a namespace for inspectors.
+    # It requires the RouteInspector to provide route inspection capabilities.
+    require "securial/inspectors/route_inspector"
+  end
+end
