--- lib/securial/inspectors/_index.rb
@@ -1,1 +1,0 @@
-require_relative "route_inspector"
