+++ lib/securial/middlewares.rb
@@ -1,0 +1,10 @@
+require_relative "middlewares/transform_request_keys"
+require_relative "middlewares/transform_response_keys"
+require_relative "middlewares/request_logger_tag"
+
+module Securial
+  module Middleware
+    # This module serves as a namespace for middlewares.
+    # It requires the necessary middleware files to provide functionality.
+  end
+end
