+++ lib/securial/middlewares/request_logger_tag.rb
