+++ lib/securial/middlewares/transform_request_keys.rb
