+++ lib/securial/middlewares/transform_response_keys.rb
